+++ src/main/java/com/pelizzaris/sufs/domain/model/Usuario.java
@@ -1,7 +1,6 @@
 package com.pelizzaris.sufs.domain.model;
 
 import jakarta.persistence.*;
-import jakarta.validation.constraints.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -17,11 +16,14 @@ public class Usuario {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private  Integer id;
+    private  Long id;
+
     @Column(name = "nome_usuario", length = 150, nullable = false)
     private String nomeUsuario;
+
     @Column(name = "email_usuario", unique = true, length = 150, nullable = false)
     private String emailUsuario;
+
     @Column(name = "status_usuario", nullable = false)
     private Boolean statusUsuario;
 }
